pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chatguys"
version = "0.1.0"
description = "A flexible multi-agent chatbot framework"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=0.19.0",
    "rich>=10.0.0",
    "prompt_toolkit>=3.0.0",
    "pyyaml>=6.0.1"
]

[project.scripts]
chatguys = "chatguys.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["chatguys"] 